Move `eth_*` files into their own directory
