+++ client/src/tests/eth/getBlockTransactionCountByHash.ts
@@ -1,7 +1,7 @@
-import { blockchain, wallet } from "../tests";
+import { blockchain, wallet } from "../../tests";
 import assert from "assert";
 
-describe("getBlockTransactionCountByNumber", () => {
+describe("getBlockTransactionCountByHash", () => {
     it("returns zero for empty block", async () => {
         if (!blockchain || !wallet) {
             throw "not ready";
@@ -11,10 +11,14 @@ describe("getBlockTransactionCountByNumber", () => {
 
         await blockchain.send("evm_mine", [{ blocks: 1 }]);
 
-        const count = await wallet.send(
-            "eth_getBlockTransactionCountByNumber",
-            [blockNumber + 1]
-        );
+        const blockHash = (await blockchain.getBlock(blockNumber + 1))?.hash;
+        if (!blockHash) {
+            throw "no block hash";
+        }
+
+        const count = await wallet.send("eth_getBlockTransactionCountByHash", [
+            blockHash,
+        ]);
         if (!count) {
             throw "no block";
         }
@@ -55,28 +59,34 @@ describe("getBlockTransactionCountByNumber", () => {
         await blockchain.send("evm_mine", [{ blocks: 1 }]);
         const mined = await response.wait(1);
 
+        const blockHash0 = (await blockchain.getBlock(blockNumber + 1))?.hash;
+        if (!blockHash0) {
+            throw "no block hash";
+        }
+
+        const blockHash1 = (await blockchain.getBlock(blockNumber + 2))?.hash;
+        if (!blockHash1) {
+            throw "no block hash";
+        }
+
         assert.equal(
-            mined?.blockNumber,
-            blockNumber + 2,
-            `transaction block (${mined?.blockNumber}) matches mined block (${
-                blockNumber + 2
-            })`
+            mined?.blockHash,
+            blockHash1,
+            `transaction block (${mined?.blockHash}) matches mined block (${blockHash1})`
         );
 
-        const count0 = await wallet.send(
-            "eth_getBlockTransactionCountByNumber",
-            [blockNumber + 1]
-        );
+        const count0 = await wallet.send("eth_getBlockTransactionCountByHash", [
+            blockHash0,
+        ]);
         if (!count0) {
             throw "no block";
         }
 
         assert.equal(parseInt(count0, 16), 1);
 
-        const count1 = await wallet.send(
-            "eth_getBlockTransactionCountByNumber",
-            [blockNumber + 2]
-        );
+        const count1 = await wallet.send("eth_getBlockTransactionCountByHash", [
+            blockHash1,
+        ]);
         if (!count1) {
             throw "no block";
         }
@@ -88,12 +98,9 @@ describe("getBlockTransactionCountByNumber", () => {
             throw "not ready";
         }
 
-        const count = await wallet.send(
-            "eth_getBlockTransactionCountByNumber",
-            [
-                "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
-            ]
-        );
+        const count = await wallet.send("eth_getBlockTransactionCountByHash", [
+            "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+        ]);
 
         assert.equal(count, null);
     });
